Hoist the MFP pushdown context out of text rendering
